PS-1-Google Lens Pro Max/requirements.txt
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.39.0
torch==2.4.1
transformers==4.45.2
